requirements.txt
matplotlib=3.7.1
numpy=1.24.2
pandas=2.0.2
pytorch=2.0.0
tqdm=4.65.0
